feat: createAlarmFromJob utility
